requirements.txt
# SimpliDOTS Slack Bot Requirements

# Core dependencies
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
beautifulsoup4>=4.9.0

# Slack integration
slack_sdk>=3.19.0

# Web framework (choose one)
flask>=2.0.0
# OR
# fastapi>=0.68.0
# uvicorn>=0.15.0

# Optional: For advanced features
pandas>=1.3.0
numpy>=1.21.0

# Development dependencies (optional)
pytest>=6.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0
